refactor(systemconfig): move internal/chaosengine/systemconfig → platform/systemconfig
