+++ aegislab/src/internal/chaosengine/guidedcli/builders.go
@@ -7,7 +7,7 @@ import (
 
 	"aegis/internal/chaosengine/handler"
 	"aegis/internal/chaosengine/resourcelookup"
-	"aegis/internal/chaosengine/systemconfig"
+	"aegis/platform/systemconfig"
 )
 
 func buildPodKill(ctx context.Context, cfg GuidedConfig, systemType systemconfig.SystemType) (map[string]any, error) {
