+++ aegislab/src/internal/chaosengine/guidedcli/k8s.go
@@ -11,7 +11,7 @@ import (
 	"k8s.io/client-go/rest"
 
 	"aegis/internal/chaosengine/resourcelookup"
-	"aegis/internal/chaosengine/systemconfig"
+	"aegis/platform/systemconfig"
 )
 
 func safeAppLabels(ctx context.Context, namespace string, systemType systemconfig.SystemType) ([]string, error) {
