+++ aegislab/src/internal/chaosengine/guidedcli/resolver.go
@@ -8,7 +8,7 @@ import (
 
 	"aegis/internal/chaosengine/handler"
 	"aegis/internal/chaosengine/resourcelookup"
-	"aegis/internal/chaosengine/systemconfig"
+	"aegis/platform/systemconfig"
 )
 
 const defaultDurationMinutes = 5
