+++ aegislab/src/internal/chaosengine/resourcelookup/db_backed_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/prometheus/client_golang/prometheus/testutil"
 
-	"aegis/internal/chaosengine/systemconfig"
+	"aegis/platform/systemconfig"
 )
 
 // stubStore returns canned chaos_points rows; mirrors the shape aegislab
