+++ aegislab/src/platform/systemconfig/registry.go
