+++ aegislab/src/platform/systemconfig/registry_test.go
